+++ RandomEncounters/EncounterGenerator.cs
@@ -1,5 +1,4 @@
-﻿using HarmonyLib;
-using System.Collections;
+﻿using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
@@ -52,6 +51,12 @@ namespace RandomEncounters
                 case int n when n > 40 && n <= 45:
                     StartCoroutine(GenerateDenseFog());
                     break;
+                case int n when n > 45 && n <= 55:
+                    StartCoroutine(GenerateFishingBonanza());
+                    break;
+                case int n when n > 55 && n <= 60:
+                    StartCoroutine(GenerateIntenseStorm());
+                    break;
             }            
         }
 
@@ -66,10 +71,11 @@ namespace RandomEncounters
         internal IEnumerator GenerateWhale()
         {
             if (!Plugin.controlSeaLifeMod.Value || Plugin.seaLifeModInstance == null) 
-                yield break;
-            
+                yield break;            
+
             for (int i = 0; i < Random.Range(1, 3); i++)
             {
+                Plugin.logger.LogDebug($"Spawning Finwhale");
                 var seaLifespawnPoint = GameState.currentBoat.position + new Vector3(Random.Range(-200, 200), -8, Random.Range(-200, 200));
                 SeaLifeMod.spawnWhale(Plugin.seaLifeModInstance, seaLifespawnPoint);
                 yield return new WaitForSeconds(0.5f);
@@ -85,7 +91,7 @@ namespace RandomEncounters
         {
             if (!Plugin.enableDenseFog.Value ||
                 DenseFog.running ||
-                Traverse.Create(WeatherStorms.instance).Method("GetNormalizedDistance").GetValue<float>() < 0.5f)
+                WeatherStorms.instance.InvokePrivateMethod<float>("GetNormalizedDistance") < 0.5f)
                 yield break;
 
             DenseFog.Spawn();
@@ -108,13 +114,95 @@ namespace RandomEncounters
             DenseFog.ClearFog();
         }
 
+        internal IEnumerator GenerateFishingBonanza()
+        {
+            if (!Plugin.enableFishingBonanza.Value)
+                yield break;
+            
+            var seagullsGO = Refs.islands[3].GetComponentsInChildren<Transform>().FirstOrDefault(t => t.name == "seagulls")?.gameObject;
+            if (seagullsGO == null)
+            {
+                Plugin.logger.LogDebug("No seagulls found");
+                yield break;
+            }
+
+            Plugin.logger.LogDebug("Starting fishing bonanza");
+            var seagulls = Instantiate(seagullsGO, Refs.shiftingWorld);
+            seagulls.GetComponent<AudioSource>().PlayOneShot(seagulls.GetComponent<AudioSource>().clip);
+            FishingBonanza.bonanzaActive = true;
+            for (int t = 0; t < Plugin.fishingBonanzaDuration.Value; t++)
+            {
+                seagulls.transform.position = GameState.currentBoat.position + GameState.currentBoat.up * 80f;
+                yield return new WaitForSeconds(1f);
+            }
+
+            Plugin.logger.LogDebug("Stopping fishing bonanza");
+            FishingBonanza.bonanzaActive = false;
+            Destroy(seagulls);
+        }
+
+        internal IEnumerator GenerateIntenseStorm()
+        {
+            if (!Plugin.enableIntenseStorm.Value)
+                yield break;
+
+            var weatherStorms = WeatherStorms.instance;
+            var storm = weatherStorms.GetCurrentStorm();
+            var lightning = storm.transform.GetChild(3).GetComponent<WanderingStormLightning>();
+            var targetRegion = RegionBlender.instance.GetPrivateField<Region>("currentTargetRegion");
+
+            var origInertiaWindScale = IntenseStorm.oceanUpdaterCrest.inertiaWindScale;
+            var origWindSpeedMult = IntenseStorm.oceanUpdaterCrest.GetPrivateField<float>("windSpeedMult");
+            var origSmallWavesMult = IntenseStorm.oceanUpdaterCrest.GetPrivateField<float>("smallWavesMult");
+            var origLightningInterval = lightning.GetPrivateField<float>("lightningInterval");
+            var origRainDensity = targetRegion.stormWeather.particles.rainDensity;
+
+            lightning.SetPrivateField("lightningInterval", 5f);
+            targetRegion.stormWeather.particles.rainDensity = 70f;
+
+            var stormDist = Vector3.Distance(Camera.main.transform.position, storm.transform.position);
+            Vector3 vector = Camera.main.transform.position - storm.transform.position;
+            vector.y = 0f;
+
+            Plugin.logger.LogDebug($"{storm.name} approaching");
+            while (stormDist > 1500f)
+            {                
+                vector = Camera.main.transform.position - storm.transform.position;                
+                vector.y = 0f;
+                Wind.currentBaseWind = vector * 50f;                
+                var translateSpeed = weatherStorms.InvokePrivateMethod<float>("GetNormalizedDistance") < weatherStorms.GetPrivateField<float>("rainBorder") ? 0.005f : 0.25f; 
+                storm.transform.Translate(vector * translateSpeed);
+                yield return new WaitForSeconds(0.3f);
+                stormDist = Vector3.Distance(Camera.main.transform.position, storm.transform.position);
+            }
+
+            Plugin.logger.LogDebug($"{storm.name} arrived");
+            IntenseStorm.oceanUpdaterCrest.inertiaWindScale = 0.22f;
+            IntenseStorm.oceanUpdaterCrest.SetPrivateField("windSpeedMult", 5f);
+            IntenseStorm.oceanUpdaterCrest.SetPrivateField("smallWavesMult", 0.4f);
+            for (int i = 0; i < Plugin.intenseStormDuration.Value; i++)
+            {               
+                Wind.currentBaseWind = vector * 50f;
+                yield return new WaitForSeconds(1f);
+            }
+
+            Plugin.logger.LogDebug($"{storm.name} dying down");
+            lightning.SetPrivateField("lightningInterval", origLightningInterval);
+            Weather.instance.currentRegion.stormWeather.particles.rainDensity = origRainDensity;
+            IntenseStorm.oceanUpdaterCrest.inertiaWindScale = origInertiaWindScale;
+            IntenseStorm.oceanUpdaterCrest.SetPrivateField("windSpeedMult", origWindSpeedMult);
+            IntenseStorm.oceanUpdaterCrest.SetPrivateField("smallWavesMult", origSmallWavesMult);            
+        }
+
+        /*
         // for testing
-        //void Update()
-        //{
-        //    if (Input.GetKeyDown(KeyCode.P))
-        //    {
-        //        Generate();
-        //    }
-        //}
+        void Update()
+        {
+            if (Input.GetKeyDown(KeyCode.P))
+            {
+                Generate();
+            }
+        }
+        */
     }
 }